+++ doc/python_reference.rst
@@ -238,7 +238,7 @@ Visualization
    plot_evoked_field
    plot_evoked_white
    plot_filter
-   plot_glass_brain
+   plot_stc_glass_brain
    plot_head_positions
    plot_ideal_filter
    plot_compare_evokeds
